requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
djangorestframework==3.16.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
